Rename block components to shared
